Update: コードのTypeScript化（途中）
