+++ src/@types/CategoryMapping.d.ts
@@ -1,0 +1,6 @@
+interface CategoryMapping {
+  start: number;
+  end?: number;
+  default: string;
+  compatible: string[];
+}
