+++ src/dict/DynamicDictionaries.ts
@@ -17,94 +17,95 @@
 
 "use strict";
 
-import doublearray from "doublearray.ts"
+import doublearray from "doublearray.ts";
 
-var TokenInfoDictionary = require("./TokenInfoDictionary");
-var ConnectionCosts = require("./ConnectionCosts");
-var UnknownDictionary = require("./UnknownDictionary");
+import TokenInfoDictionary from "./TokenInfoDictionary";
+import ConnectionCosts from "./ConnectionCosts";
+import UnknownDictionary from "./UnknownDictionary";
+import DoubleArray from "doublearray.ts/dist/doubleArrayClass";
+import { ArrayBuffer } from "doublearray.ts/dist/types";
 
-export interface DynamicDictionaries {
-    trie: 
+class DynamicDictionaries {
+  trie: DoubleArray;
+  token_info_dictionary: TokenInfoDictionary;
+  connection_costs: ConnectionCosts;
+  unknown_dictionary: UnknownDictionary;
+
+  /**
+   * Dictionaries container for Tokenizer
+   * @param {DoubleArray} trie
+   * @param {TokenInfoDictionary} token_info_dictionary
+   * @param {ConnectionCosts} connection_costs
+   * @param {UnknownDictionary} unknown_dictionary
+   * @constructor
+   */
+  constructor(
+    trie?: DoubleArray | null,
+    token_info_dictionary?: TokenInfoDictionary | null,
+    connection_costs?: ConnectionCosts | null,
+    unknown_dictionary?: UnknownDictionary | null,
+  ) {
+    if (trie != null) {
+      this.trie = trie;
+    } else {
+      this.trie = doublearray.builder(0).build([
+        { k: "", v: 1 },
+      ]);
+    }
+    if (token_info_dictionary != null) {
+      this.token_info_dictionary = token_info_dictionary;
+    } else {
+      this.token_info_dictionary = new TokenInfoDictionary();
+    }
+    if (connection_costs != null) {
+      this.connection_costs = connection_costs;
+    } else {
+      // backward_size * backward_size
+      this.connection_costs = new ConnectionCosts(0, 0);
+    }
+    if (unknown_dictionary != null) {
+      this.unknown_dictionary = unknown_dictionary;
+    } else {
+      this.unknown_dictionary = new UnknownDictionary();
+    }
+  }
+
+  // from base.dat & check.dat
+  loadTrie(base_buffer: ArrayBuffer, check_buffer: ArrayBuffer) {
+    this.trie = doublearray.load(base_buffer, check_buffer);
+  }
+
+  loadTokenInfoDictionaries(
+    token_info_buffer: Uint8Array,
+    pos_buffer: Uint8Array,
+    target_map_buffer: Uint8Array,
+  ) {
+    this.token_info_dictionary.loadDictionary(token_info_buffer);
+    this.token_info_dictionary.loadPosVector(pos_buffer);
+    this.token_info_dictionary.loadTargetMap(target_map_buffer);
+  }
+
+  loadConnectionCosts(cc_buffer: Int16Array) {
+    this.connection_costs.loadConnectionCosts(cc_buffer);
+  }
+
+  loadUnknownDictionaries(
+    unk_buffer: Uint8Array,
+    unk_pos_buffer: Uint8Array,
+    unk_map_buffer: Uint8Array,
+    cat_map_buffer: Uint8Array,
+    compat_cat_map_buffer: Uint32Array,
+    invoke_def_buffer: Uint8Array,
+  ) {
+    this.unknown_dictionary.loadUnknownDictionaries(
+      unk_buffer,
+      unk_pos_buffer,
+      unk_map_buffer,
+      cat_map_buffer,
+      compat_cat_map_buffer,
+      invoke_def_buffer,
+    );
+  }
 }
 
-/**
- * Dictionaries container for Tokenizer
- * @param {DoubleArray} trie
- * @param {TokenInfoDictionary} token_info_dictionary
- * @param {ConnectionCosts} connection_costs
- * @param {UnknownDictionary} unknown_dictionary
- * @constructor
- */
-function DynamicDictionaries(
-  trie,
-  token_info_dictionary,
-  connection_costs,
-  unknown_dictionary,
-) {
-  if (trie != null) {
-    this.trie = trie;
-  } else {
-    this.trie = doublearray.builder(0).build([
-      { k: "", v: 1 },
-    ]);
-  }
-  if (token_info_dictionary != null) {
-    this.token_info_dictionary = token_info_dictionary;
-  } else {
-    this.token_info_dictionary = new TokenInfoDictionary();
-  }
-  if (connection_costs != null) {
-    this.connection_costs = connection_costs;
-  } else {
-    // backward_size * backward_size
-    this.connection_costs = new ConnectionCosts(0, 0);
-  }
-  if (unknown_dictionary != null) {
-    this.unknown_dictionary = unknown_dictionary;
-  } else {
-    this.unknown_dictionary = new UnknownDictionary();
-  }
-}
-
-// from base.dat & check.dat
-DynamicDictionaries.prototype.loadTrie = function (base_buffer, check_buffer) {
-  this.trie = doublearray.load(base_buffer, check_buffer);
-  return this;
-};
-
-DynamicDictionaries.prototype.loadTokenInfoDictionaries = function (
-  token_info_buffer,
-  pos_buffer,
-  target_map_buffer,
-) {
-  this.token_info_dictionary.loadDictionary(token_info_buffer);
-  this.token_info_dictionary.loadPosVector(pos_buffer);
-  this.token_info_dictionary.loadTargetMap(target_map_buffer);
-  return this;
-};
-
-DynamicDictionaries.prototype.loadConnectionCosts = function (cc_buffer) {
-  this.connection_costs.loadConnectionCosts(cc_buffer);
-  return this;
-};
-
-DynamicDictionaries.prototype.loadUnknownDictionaries = function (
-  unk_buffer,
-  unk_pos_buffer,
-  unk_map_buffer,
-  cat_map_buffer,
-  compat_cat_map_buffer,
-  invoke_def_buffer,
-) {
-  this.unknown_dictionary.loadUnknownDictionaries(
-    unk_buffer,
-    unk_pos_buffer,
-    unk_map_buffer,
-    cat_map_buffer,
-    compat_cat_map_buffer,
-    invoke_def_buffer,
-  );
-  return this;
-};
-
-module.exports = DynamicDictionaries;
+export default DynamicDictionaries;
